+++ org.eclipse.winery.repository.rest/src/main/java/org/eclipse/winery/repository/rest/resources/entitytemplates/artifacttemplates/FilesResource.java
@@ -22,6 +22,7 @@ import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.HeaderParam;
 import javax.ws.rs.POST;
+import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
 import javax.ws.rs.Produces;
@@ -142,6 +143,17 @@ public class FilesResource {
 	@Path("/{fileName}")
 	@Consumes(MediaType.APPLICATION_JSON)
 	public Response postFile(@PathParam("fileName") String fileName, ArtifactResourceApiData data) {
+		if (StringUtils.isEmpty(fileName)) {
+			return Response.status(Status.BAD_REQUEST).build();
+		}
+		RepositoryFileReference ref = this.fileName2fileRef(fileName, false);
+		return RestUtils.putContentToFile(ref, data.content, MediaType.TEXT_PLAIN_TYPE);
+	}
+	
+	@PUT
+	@Path("/{fileName}")
+	@Consumes(MediaType.APPLICATION_JSON)
+	public Response putFile(@PathParam("fileName") String fileName, ArtifactResourceApiData data) {
 		if (StringUtils.isEmpty(fileName)) {
 			return Response.status(Status.BAD_REQUEST).build();
 		}