+++ org.eclipse.winery.repository.ui/src/app/instance/instance.service.ts
@@ -13,7 +13,6 @@
 import { Injectable } from '@angular/core';
 import { Headers, Http, RequestOptions } from '@angular/http';
 import { Observable } from 'rxjs';
-import { isNullOrUndefined } from 'util';
 import { backendBaseURL } from '../configuration';
 import { WineryInstance, WineryTopologyTemplate } from '../wineryInterfaces/wineryComponent';
 import { ToscaComponent } from '../wineryInterfaces/toscaComponent';
@@ -57,7 +56,7 @@ export class InstanceService {
                 subMenu = ['Properties Definition', 'Inheritance', 'Documentation', 'XML'];
                 break;
             case ToscaTypes.ArtifactTemplate:
-                subMenu = ['Files', 'Source', 'Properties', 'Documentation', 'XML'];
+                subMenu = ['README', 'Files', 'Source', 'Properties', 'Documentation', 'XML'];
                 break;
             case ToscaTypes.RequirementType:
                 subMenu = ['Required Capability Type', 'Properties Definition', 'Inheritance', 'Documentation', 'XML'];
@@ -112,15 +111,15 @@ export class InstanceService {
     }
 
     public getComponentData(): Observable<WineryInstance> {
-        const headers = new Headers({'Content-Type': 'application/xml'});
-        const options = new RequestOptions({headers: headers});
+        const headers = new Headers({ 'Content-Type': 'application/xml' });
+        const options = new RequestOptions({ headers: headers });
         return this.http.get(backendBaseURL + this.path + '/', options)
             .map(res => res.json());
     }
 
     private getTopologyTemplate(): Observable<WineryTopologyTemplate> {
-        const headers = new Headers({'Content-Type': 'application/json'});
-        const options = new RequestOptions({headers: headers});
+        const headers = new Headers({ 'Content-Type': 'application/json' });
+        const options = new RequestOptions({ headers: headers });
         return this.http.get(backendBaseURL + this.path + '/topologytemplate/', options)
             .map(res => res.json());
     }